+++ src/siphash.h
@@ -15,5 +15,5 @@
     HALF_ROUND(v2,v1,v0,v3,17,21);        \
     HALF_ROUND(v0,v1,v2,v3,13,16);        \
     HALF_ROUND(v2,v1,v0,v3,17,21);
-uint64_t siphash24(const void* src, unsigned long src_sz, const char key[16]);
-uint64_t siphash24_keyed(const void* src, unsigned long src_sz);
+uint64_t siphash24(const void* src, unsigned src_sz, const char key[16]);
+uint64_t siphash24_keyed(const void* src, unsigned src_sz);
